Move utils and models to base folder
